为协议ID定义类型，将协议ID长度调整为4字节; 协议注册标不再接受指定协议ID, 由注册标自动根据协议名称哈西; 协议注册表不再提供通过协议名称创建消息对象以及通过协议ID获取协议名称等功能

+++ infra/actor/codec.go
@@ -79,7 +79,7 @@ func (c *Codec) Encode(allocator gactor.PacketAllocator, payload any) ([]byte, e
 		if err := allocator.AllocBuf(&buffer, 2+len(msgBytes)); err != nil {
 			return nil, pkgerrors.WithMessagef(err, "alloc buf failed")
 		}
-		if err := buffer.WriteUint16(p.PID); err != nil {
+		if err := buffer.WriteUint32(uint32(p.PID)); err != nil {
 			return nil, pkgerrors.WithMessage(err, "write pid to buf failed")
 		}
 		if _, err := buffer.Write(msgBytes); err != nil {
